nlp-service/requirements.txt
fastapi
uvicorn
spacy
scikit-learn
sentence-transformers
faiss-cpu
pdfplumber
python-docx
PyMuPDF
python-multipart
yake
numpy
